+++ helpers/dev.js
@@ -420,7 +420,7 @@ export const shouldOutputForLogLevel = (logType: string): boolean => {
 export const shouldOutputForFunctionName = (pluginInfo: any): boolean => {
   const pluginSettings = typeof DataStore !== 'undefined' ? DataStore.settings : null
   if (pluginSettings && pluginSettings.hasOwnProperty('_logFunctionRE')) {
-    const functionRE = new RegExp(pluginSettings['_logFunctionRE'])
+    const functionRE = new RegExp(pluginSettings['_logFunctionRE'], 'i')
     const infoStr: string = pluginInfo === 'object' ? pluginInfo['plugin.id'] : String(pluginInfo)
     return functionRE.test(infoStr)
   }